Move RTLD_NOW from compile-time to runtime default, so it becomes runtime-overridable
This change moves the RTLD_NOW default from compile-time in C into runtime in Python.

This is useful to allow ctypes to lazy-bind at dlopen occasionally,
i.e. for the special purpose of inspecting genuine dylib's under darling.

This change is fully backward-compatible : people who are not aware of it,
won't notice any difference - RTLD_NOW remains the default, and
ctypes defaults to try to resolve all symbols at load. What it does, is
that one can now explicitly ask for RTLD_LAZY at the python level, and
have this flag passed along directly to dlopen. This exhancement is useful
when not all symbols are resolvable, e. g. under darling when examining
genuine dylib's.